Kernel: Make sure MSB is not set on SMO keys

+++ kernel/include/kernel/Memory/SharedMemoryObject.h
@@ -14,7 +14,7 @@ namespace Kernel
 	class SharedMemoryObjectManager
 	{
 	public:
-		using Key = uint32_t;
+		using Key = size_t;
 
 	public:
 		static BAN::ErrorOr<void> initialize();
